feat(cli): cloudflare/aws namespaces + cloudflare state logs
Move bootstrap under per-cloud namespaces (alchemy aws bootstrap,
alchemy cloudflare bootstrap). Add alchemy cloudflare state logs
[--tail] to debug the alchemy-state-store worker, reusing the worker
tail/query helpers from Cloudflare/Logs.
