+++ packages/alchemy/src/Cli/main.ts
@@ -11,7 +11,8 @@ import packageJson from "../../package.json" with { type: "json" };
 
 import { checkLatestVersion } from "./checkVersion.ts";
 import { handleCancellation } from "./commands/_shared.ts";
-import { bootstrapCommand } from "./commands/bootstrap.ts";
+import { awsCommand } from "./commands/aws.ts";
+import { cloudflareCommand } from "./commands/cloudflare.ts";
 import {
   deployCommand,
   destroyCommand,
@@ -27,7 +28,8 @@ import { inkCLI } from "./tui/InkCLI.tsx";
 
 const root = Command.make("alchemy", {}).pipe(
   Command.withSubcommands([
-    bootstrapCommand,
+    awsCommand,
+    cloudflareCommand,
     deployCommand,
     devCommand,
     destroyCommand,
